SourceGen: 1) Rename find_injectable_modules.dart to compile_type.dart. 2) Use the CompileTypeMetadataVisitor in ComponentVisitor and PipeVistor. 3) Fix NormalizedComponentVisitor to only find Components, and not Directives.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=139924146
